+++ examples/bert/yaml/pre_process/component.yaml
@@ -8,7 +8,7 @@ outputs:
  
 implementation:
   container:
-    image: public.ecr.aws/y1x1p2u5/pytorch_pipeline:21052021_v1.0
+    image: arvindideas2it/pytorch_kfp_components:r0.6
     command: ["python3", "examples/bert/bert_pre_process.py"] 
     args:
     - --output_path
