+++ examples/bert/yaml/train/component.yaml
@@ -14,7 +14,7 @@ outputs:
   - {name: MLPipeline Metrics}
 implementation:
   container:
-    image: public.ecr.aws/y1x1p2u5/pytorch_pipeline:21052021_v1.0
+    image: arvindideas2it/pytorch_kfp_components:r0.6
     command: ["python3", "examples/bert/agnews_classification_pytorch.py"]
     args:
       - --dataset_path
