+++ examples/cifar10/yaml/train/component.yaml
@@ -16,7 +16,7 @@ outputs:
   
 implementation:
   container:
-    image: public.ecr.aws/y1x1p2u5/pytorch_pipeline:21052021_v1.0
+    image: arvindideas2it/pytorch_kfp_components:r0.6
     command: ["python3", "examples/cifar10/cifar10_pytorch.py"]
     args:
     - --dataset_path
