+++ examples/common/minio/component.yaml
@@ -12,8 +12,8 @@ outputs:
 
 implementation:
   container:
-    image: public.ecr.aws/y1x1p2u5/pytorch_pipeline:21052021_v1.0
-    command: ["python3", "pytorch_pipeline/components/minio/upload_to_minio.py"]
+    image: arvindideas2it/pytorch_kfp_components:r0.6
+    command: ["python3", "pytorch_kfp_components/components/minio/upload_to_minio.py"]
     args:
       - --bucket_name
       - {inputValue: bucket_name}
