+++ pytorch_kfp_components/components/mar/component.py
@@ -6,9 +6,9 @@
 # LICENSE file in the root directory of this source tree.
 """Mar Generation Component Class."""
 
-from pytorch_pipeline.components.mar.executor import Executor
-from pytorch_pipeline.types import standard_component_specs
-from pytorch_pipeline.components.base.base_component import BaseComponent
+from pytorch_kfp_components.components.mar.executor import Executor
+from pytorch_kfp_components.types import standard_component_specs
+from pytorch_kfp_components.components.base.base_component import BaseComponent
 
 
 class MarGeneration(BaseComponent):  #pylint: disable=R0903
