+++ pytorch_kfp_components/components/mar/executor.py
@@ -12,8 +12,8 @@ import tempfile
 import subprocess
 from pathlib import Path
 import wget
-from pytorch_pipeline.components.base.base_executor import BaseExecutor
-from pytorch_pipeline.types import standard_component_specs
+from pytorch_kfp_components.components.base.base_executor import BaseExecutor
+from pytorch_kfp_components.types import standard_component_specs
 
 
 class Executor(BaseExecutor):
